# api-docs/swagger.json
{
  "info": {
    "title": "Staffing API",
    "description": "Staffing API Sandbox",
    "version": "1.0.0"
  },
  "schemes": [
    "https"
  ],
  "basePath": "/api",
  "swagger": "2.0",
  "paths": {
    "/businesses": {
      "get": {
        "tags": [
          "business"
        ],
        "summary": "List businesses by query",
        "description": "List businesses by query. if no query is sent, list all businesses",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation.",
            "schema": {
              "$ref": "#/definitions/Business"
            }
          },
          "401": {
            "description": "unauthorized"
          },
          "500": {
            "description": "internal error"
          }
        }
      }
    },
...
    "/companies": {
      "get": {
        "tags": [
          "Company"
        ],
        "summary": "List companies by query",
        "description": "List companies by query. if no query is sent, list all companies",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation.",
            "schema": {
              "$ref": "#/definitions/Company"
            }
          },
          "401": {
            "description": "unauthorized"
          },
          "500": {
            "description": "internal error"
          }
        }
      },
      "put": {
        "tags": [
          "Company"
        ],
        "summary": "Update company info",
        "description": "Update company info and store in edit layer without changing original info",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "_id",
            "type": "string",
            "description": "_id of company document",
            "required": true
          },
          {
            "name": "source",
            "type": "string",
            "description": "Who is editing. It will be user when editing from client forms. Or it can by any scrape source like crunchbase, craft etc",
            "required": true
          },
          {
            "name": "companyName",
            "type": "string"
          },
          {
            "name": "description",
            "type": "string"
          },
          {
            "name": "foundersNames",
            "type": "array",
            "items": {
              "type": "string"
            },
            "in": "formData"
          },
          {
            "name": "categories",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "keyPeople",
            "type": "array",
            "items": {
              "type": "object",
              "schema": {
                "$ref": "#/definitions/KeyPeople"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation.",
            "schema": {
              "$ref": "#/definitions/Company"
            }
          },
          "401": {
            "description": "unauthorized"
          },
          "500": {
            "description": "internal error"
          }
        }
      }
    },
...
    "/jobs": {
      "get": {
        "tags": [
          "Job"
        ],
        "summary": "List jobs by query",
        "description": "List jobs by query. if no query is sent, list all jobs",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation.",
            "schema": {
              "$ref": "#/definitions/Job"
            }
          },
          "401": {
            "description": "unauthorized"
          },
          "500": {
            "description": "internal error"
          }
        }
      }
    },
...
    "/scrape": {
      "get": {
        "tags": [
          "Scrape"
        ],
        "summary": "Scrape a web site",
        "description": "Scrape a web site by company type.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "types",
            "in": "formData",
            "description": "scrape type which also represent company",
            "required": true,
            "type": "integer"
          },
          {
            "name": "terms",
            "in": "formData",
            "description": "Search terms.",
            "required": true,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "formData",
            "description": "limit how many records to scrap",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation."
          },
          "400": {
            "description": "error fetching"
          },
          "401": {
            "description": "unauthorized"
          },
          "404": {
            "description": "not found"
          },
          "500": {
            "description": "internal error"
          }
        }
      }
    },
    "/scrape/companies": {
      "get": {
        "tags": [
          "Scrape companies"
        ],
        "summary": "Scrape companies from the specified source.",
        "description": "Scrape companies from the specified source, crunchBase by default.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "request.query",
            "description": "companies category/market/industry we want to scrape, default to all",
            "required": false,
            "type": "string"
          },
          {
            "name": "source",
            "in": "request.query",
            "description": "website we want to do the scraping from. default to crunchBase",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "request.query",
            "description": "limit how many records to scrape, default to 1000",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation."
          },
          "400": {
            "description": "error fetching"
          },
          "401": {
            "description": "unauthorized"
          },
          "404": {
            "description": "not found"
          },
          "500": {
            "description": "internal error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "OpeningHours": {
      "type": "object",
      "properties": {
        "mon": {
          "type": "string"
        },
        "tue": {
          "type": "string"
        },
        "wed": {
          "type": "string"
        },
        "thu": {
          "type": "string"
        },
        "fri": {
          "type": "string"
        },
        "sat": {
          "type": "string"
        },
        "sun": {
          "type": "string"
        }
      }
    },
    "Rating": {
      "type": "object",
      "properties": {
        "five": {
          "type": "string"
        },
        "four": {
          "type": "string"
        },
        "three": {
          "type": "string"
        },
        "two": {
          "type": "string"
        },
        "one": {
          "type": "string"
        }
      }
    },
    "keyValue": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "Business": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "website": {
          "type": "string"
        },
        "owners": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "street": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "postalCode": {
          "type": "string"
        },
        "neighborhood": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "categories": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "coordinates": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "partOfChain": {
          "type": "boolean"
        },
        "facebook": {
          "type": "string"
        },
        "twitter": {
          "type": "string"
        },
        "yelpUrl": {
          "type": "string"
        },
        "yelpAvatar": {
          "type": "string"
        },
        "priceRange": {
          "type": "string"
        },
        "priceDescription": {
          "type": "string"
        },
        "healthInspection": {
          "type": "string"
        },
        "hours": {
          "type": "object",
          "schema": {
            "$ref": "#/definitions/OpeningHours"
          }
        },
        "reviews": {
          "type": "object",
          "schema": {
            "$ref": "#/definitions/Rating"
          }
        },
        "yelpingSince": {
          "type": "string"
        },
        "additionalInfo": {
          "type": "array",
          "items": {
            "type": "object",
            "schema": {
              "$ref": "#/definitions/keyValue"
            }
          }
        },
        "source": {
          "type": "string"
        }
      }
    },
    "Company": {
      "type": "object",
      "properties": {
        "companyName": {
          "type": "string"
        },
        "foundersNames": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "cityHeadQuartersIn": {
          "type": "string"
        },
        "numberOfEmployees": {
          "type": "number"
        },
        "websiteLink": {
          "type": "string"
        },
        "fundingRounds": {
          "type": "number"
        },
        "fundingAmount": {
          "type": "string"
        },
        "lastFundingDate": {
          "type": "string"
        },
        "dateFounded": {
          "type": "string"
        },
        "categories": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "angelId": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "acquisitionsNumber": {
          "type": "string"
        },
        "linkedInURL": {
          "type": "string"
        },
        "twitterURL": {
          "type": "string"
        },
        "facebookURL": {
          "type": "string"
        },
        "youtubeURL": {
          "type": "string"
        },
        "instagramURL": {
          "type": "string"
        },
        "angelURL": {
          "type": "string"
        },
        "crunchBaseURL": {
          "type": "string"
        },
        "craftURL": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "keyPeople": {
          "type": "array",
          "items": {
            "type": "object",
            "schema": {
              "$ref": "#/definitions/KeyPeople"
            }
          }
        },
        "source": {
          "type": "string"
        },
        "logos": {
          "type": "array",
          "items": {
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Logos"
            }
          }
        },
        "otherLocations": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "edits": {
          "type": "array",
          "items": {
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Edits"
            }
          }
        }
      }
    },
...
    "KeyPeople": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "linkedIn": {
          "type": "string"
        },
        "facebook": {
          "type": "string"
        },
        "twitter": {
          "type": "string"
        },
        "website": {
          "type": "string"
        },
        "other": {
          "type": "string"
        }
      }
    },
    "Logos": {
      "type": "object",
      "properties": {
        "craftLogo": {
          "type": "string"
        },
        "crunchBaseLogo": {
          "type": "string"
        }
      }
    },
    "Edits": {
      "type": "object",
      "properties": {
        "companyName": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "numberOfEmployees": {
          "type": "number"
        },
        "dateFounded": {
          "type": "string"
        },
        "websiteLink": {
          "type": "string"
        },
        "linkedInURL": {
          "type": "string"
        },
        "twitterURL": {
          "type": "string"
        },
        "facebookURL": {
          "type": "string"
        },
        "fundingAmount": {
          "type": "string"
        },
        "fundingRounds": {
          "type": "number"
        },
        "lastRound": {
          "type": "string"
        },
        "foundersNames": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "categories": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "keyPeople": {
          "type": "array",
          "items": {
            "type": "object",
            "schema": {
              "$ref": "#/definitions/KeyPeople"
            }
          }
        },
        "source": {
          "type": "string"
        },
        "revenue": {
          "type": "string"
        }
      }
    },
...
    "Job": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "jobLink": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "company": {
          "type": "string"
        },
        "companyLink": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "date": {
          "type": "string"
        }
      }
    },
...
  },
...
  "securityDefinitions": {}
}